# swagger_output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EGI Media AI Backend API",
    "version": "0.1.0",
    "description": "API contract for the EGI Media AI decision-support service. The service is tenant- and company-scoped, reads editorial articles from egi-media-backend, and exposes only validated AI projections. It never exposes prompts, provider credentials, raw provider output, or CMS write operations."
  },
  "servers": [
    {
      "url": "http://localhost:5003",
      "description": "Local AI backend"
    }
  ],
...
  "paths": {
...
    "/api/v1/internal/issues/{issueId}/analyze": { "post": { "tags": ["Internal Issue Analysis"], "summary": "Run T07 issue analysis", "description": "Analyzes all linked, published, fresh article evidence for one active issue.", "security": [{ "internalServiceAuth": [] }], "parameters": [{ "name": "issueId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } }, { "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }], "responses": { "200": { "description": "Validated issue analysis" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Conflict" }, "502": { "$ref": "#/components/responses/ProviderError" } } } },
    "/api/v1/internal/analyses/{analysisId}/labels": { "post": { "tags": ["Internal Issue Analysis"], "summary": "Run T08 claim labels", "description": "Labels every existing T07 claim exactly once; it cannot add or rewrite claims.", "security": [{ "internalServiceAuth": [] }], "parameters": [{ "name": "analysisId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } }, { "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }], "responses": { "200": { "description": "Validated claim labels" }, "403": { "$ref": "#/components/responses/Forbidden" }, "502": { "$ref": "#/components/responses/ProviderError" } } } },
    "/api/v1/internal/analyses/{analysisId}/promote-current": { "post": { "tags": ["Internal Issue Analysis"], "summary": "Promote citation-gated current analysis", "description": "Revalidates fresh canonical evidence, citations, and T08 labels before replacing the current analysis.", "security": [{ "internalServiceAuth": [] }], "parameters": [{ "name": "analysisId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } }, { "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }], "responses": { "200": { "description": "Current validated analysis" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Conflict" }, "502": { "$ref": "#/components/responses/ProviderError" } } } },
...
    "/api/v1/companies/{companyId}/alert-preference": { "put": { "tags": ["Alert Eligibility"], "summary": "Upsert alert preference", "description": "Stores recipient preference used by backend alert eligibility rules. It does not send email.", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/CompanyIdPath" }, { "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AlertPreferenceRequest" } } } }, "responses": { "200": { "description": "Alert preference saved", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AlertPreferenceResponse" } } } }, "400": { "$ref": "#/components/responses/ValidationError" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } } } },
    "/api/v1/internal/alerts/eligibility": { "post": { "tags": ["Alert Eligibility"], "summary": "Evaluate alert eligibility", "description": "Rules-only decision for preference, priority, dedupe, and quiet hours. Material update with no product decision is fail-closed as suppressed. This endpoint never sends email and never exposes recipient or subject selection to an AI model.", "security": [{ "internalServiceAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AlertEligibilityRequest" } } } }, "responses": { "200": { "description": "Rules-only eligibility decision", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AlertEligibilityResponse" } } } }, "400": { "$ref": "#/components/responses/ValidationError" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "422": { "$ref": "#/components/responses/BusinessRuleError" }, "503": { "$ref": "#/components/responses/ServiceUnavailable" } } } },
...
    "/api/v1/internal/alerts/{alertEventId}/direct-blurb": { "post": { "tags": ["Direct Alert"], "summary": "Generate T12 direct-alert blurb", "description": "Runs Nano only after the backend has produced an eligible direct alert event. The model receives validated issue evidence and claim references, never recipient, subject, email address, or delivery instructions. This endpoint generates content only and never sends email.", "security": [{ "internalServiceAuth": [] }], "parameters": [{ "name": "alertEventId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } }, { "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }], "requestBody": { "required": false, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false, "maxProperties": 0 } } } }, "responses": { "200": { "description": "Validated T12 direct-alert blurb", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DirectBlurbResponse" } } } }, "400": { "$ref": "#/components/responses/ValidationError" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "422": { "$ref": "#/components/responses/BusinessRuleError" }, "502": { "$ref": "#/components/responses/ProviderError" } } } },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "End-user identity. Tenant/company scope is derived and verified server-side." },
      "internalServiceAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "Service JWT", "description": "Internal service or authorized operator identity; never available to frontend." }
    },
    "parameters": {
...
      "CompanyIdPath": { "name": "companyId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } },
...
      "IdempotencyKeyHeader": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 16, "maxLength": 255 } },
...
      "RequestIdHeader": { "name": "X-Request-Id", "in": "header", "required": false, "schema": { "type": "string", "maxLength": 100 } }
    },
    "responses": {
      "Unauthorized": { "description": "Authentication missing or invalid", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Forbidden": { "description": "Caller lacks required tenant/company/role permission", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
...
      "Conflict": { "description": "Idempotency key, version, or state conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "ValidationError": { "description": "Request schema validation failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "BusinessRuleError": { "description": "Valid request cannot proceed because of pipeline/state/delivery business rules", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "ProviderError": { "description": "External AI/provider request failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "ServiceUnavailable": { "description": "Required backend dependency is unavailable", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
    },
    "schemas": {
      "Uuid": { "type": "string", "format": "uuid" },
      "Timestamp": { "type": "string", "format": "date-time" },
...
      "ErrorCode": { "type": "string", "enum": ["VALIDATION_ERROR", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "VERSION_CONFLICT", "IDEMPOTENCY_CONFLICT", "BUSINESS_RULE_FAILED", "AI_OUTPUT_INVALID", "SOURCE_VALIDATION_FAILED", "PIPELINE_PENDING", "INTERNAL_ERROR"] },
      "ApiMeta": {
        "type": "object",
        "required": ["request_id"],
        "properties": {
          "request_id": { "type": "string" },
          "trace_id": { "type": "string" },
          "page": { "type": "integer", "minimum": 1 },
          "limit": { "type": "integer", "minimum": 1 },
          "total": { "type": "integer", "minimum": 0 },
          "total_pages": { "type": "integer", "minimum": 0 }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "error", "meta"],
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "error": { "type": "object", "required": ["code", "message"], "properties": { "code": { "$ref": "#/components/schemas/ErrorCode" }, "message": { "type": "string" }, "fields": { "type": "array", "items": { "type": "object", "properties": { "field": { "type": "string" }, "code": { "type": "string" } } } } } },
          "meta": { "$ref": "#/components/schemas/ApiMeta" }
        }
      },
...
      "QuietHours": { "type": "object", "required": ["start", "end"], "additionalProperties": false, "properties": { "start": { "type": "string", "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$" }, "end": { "type": "string", "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$" } } },
      "AlertPreferenceRequest": { "type": "object", "required": ["recipient_id", "direct_high_enabled", "daily_digest_enabled", "timezone", "quiet_hours"], "additionalProperties": false, "properties": { "recipient_id": { "type": "string", "minLength": 1, "maxLength": 255 }, "direct_high_enabled": { "type": "boolean" }, "daily_digest_enabled": { "type": "boolean" }, "timezone": { "type": "string", "minLength": 1, "maxLength": 100 }, "quiet_hours": { "oneOf": [{ "type": "null" }, { "$ref": "#/components/schemas/QuietHours" }] } } },
      "AlertPreference": { "allOf": [{ "$ref": "#/components/schemas/AlertPreferenceRequest" }] },
      "AlertPreferenceResponse": { "type": "object", "required": ["success", "data", "meta"], "properties": { "success": { "type": "boolean", "enum": [true] }, "data": { "$ref": "#/components/schemas/AlertPreference" }, "meta": { "$ref": "#/components/schemas/ApiMeta" } } },
      "AlertEligibilityRequest": { "type": "object", "required": ["tenant_id", "company_id", "issue_id", "development_id", "recipient_id"], "additionalProperties": false, "properties": { "tenant_id": { "$ref": "#/components/schemas/Uuid" }, "company_id": { "$ref": "#/components/schemas/Uuid" }, "issue_id": { "$ref": "#/components/schemas/Uuid" }, "development_id": { "$ref": "#/components/schemas/Uuid" }, "recipient_id": { "type": "string", "minLength": 1, "maxLength": 255 } } },
      "AlertDecision": { "type": "object", "required": ["channel", "status", "reason_code"], "properties": { "channel": { "type": "string", "enum": ["langsung", "ringkasan", "none"] }, "status": { "type": "string", "enum": ["eligible", "suppressed"] }, "reason_code": { "type": "string" } } },
      "AlertEligibilityResponse": { "type": "object", "required": ["success", "data", "meta"], "properties": { "success": { "type": "boolean", "enum": [true] }, "data": { "type": "object", "required": ["decision", "email_send"], "properties": { "decision": { "$ref": "#/components/schemas/AlertDecision" }, "email_send": { "type": "boolean", "enum": [false] } } }, "meta": { "$ref": "#/components/schemas/ApiMeta" } } },
      "DirectBlurb": { "type": "object", "required": ["direct_blurb_id", "alert_event_id", "issue_id", "development_id", "new_development_blurb", "short_impact_blurb", "source_claim_ids", "prompt_version", "generated_at"], "properties": { "direct_blurb_id": { "$ref": "#/components/schemas/Uuid" }, "alert_event_id": { "$ref": "#/components/schemas/Uuid" }, "issue_id": { "$ref": "#/components/schemas/Uuid" }, "development_id": { "$ref": "#/components/schemas/Uuid" }, "new_development_blurb": { "type": "string", "maxLength": 500 }, "short_impact_blurb": { "type": "string", "maxLength": 500 }, "source_claim_ids": { "type": "array", "minItems": 1, "items": { "type": "string" } }, "prompt_version": { "type": "string" }, "generated_at": { "$ref": "#/components/schemas/Timestamp" } } },
      "DirectBlurbResponse": { "type": "object", "required": ["success", "data", "meta"], "properties": { "success": { "type": "boolean", "enum": [true] }, "data": { "type": "object", "required": ["blurb", "reused", "email_send"], "properties": { "blurb": { "$ref": "#/components/schemas/DirectBlurb" }, "reused": { "type": "boolean" }, "email_send": { "type": "boolean", "enum": [false] } } }, "meta": { "$ref": "#/components/schemas/ApiMeta" } } },
...
    }
  }
}